refactor the code to use with webpack, move every module to its own file and added jest dependency
