+++ src/index.js
@@ -1,0 +1,6 @@
+import './assets/css/style.css';
+
+import Game from './game';
+  
+Game.firstAddListener();
+
