+++ src/player.js
@@ -1,0 +1,7 @@
+const Player = (name, symbol) => {
+  const getName = () => name;
+  const getSymbol = () => symbol;
+  return { getName, getSymbol };
+};
+
+export default Player;
